+++ scripts/test.sh
@@ -1,8 +1,18 @@
 #!/bin/bash
-echo "=== FORMAT ===" &&
-cargo fmt &&
+echo "=== CHECK ===" &&
+cargo check &&
 echo "=== CLIPPY ===" &&
 cargo clippy -- -D warnings &&
-echo "=== TARPAULIN ===" &&
-cargo tarpaulin -o lcov &&
+echo "=== TEST ===" &&
+cargo test &&
+echo "=== TODOS ===" &&
+./scripts/lint_string.sh "TODO " &&
+./scripts/lint_string.sh "TODO:" &&
+./scripts/lint_string.sh "FIXME " &&
+./scripts/lint_string.sh "FIXME:" &&
+./scripts/lint_string.sh "todo " &&
+./scripts/lint_string.sh "todo:" &&
+./scripts/lint_string.sh "fixme " &&
+./scripts/lint_string.sh "fixme:" &&
+./scripts/lint_string.sh "dbg!" &&
 echo "=== SUCCESS ==="