+++ src/projects.rs
@@ -142,7 +142,7 @@ pub async fn next_task(config: Config, project: &Project) -> Result<String, Erro
     match fetch_next_task(&config, project).await {
         Ok(Some((task, remaining))) => {
             config.set_next_id(&task.id).save().await?;
-            let task_string = task.fmt(&config, FormatType::Single, false);
+            let task_string = task.fmt(&config, FormatType::Single, false, true).await?;
             Ok(format!("{task_string}\n{remaining} task(s) remaining"))
         }
         Ok(None) => Ok(color::green_string("No tasks on list")),
@@ -372,7 +372,7 @@ pub async fn schedule(
     } else {
         let mut handles = Vec::new();
         for task in filtered_tasks.iter() {
-            if let Some(handle) = tasks::spawn_schedule_task(config.clone(), task.clone())? {
+            if let Some(handle) = tasks::spawn_schedule_task(config.clone(), task.clone()).await? {
                 handles.push(handle);
             }
         }
@@ -390,7 +390,8 @@ pub async fn move_task_to_project(
     task: Task,
     sections: &[Section],
 ) -> Result<JoinHandle<()>, Error> {
-    println!("{}", task.fmt(config, FormatType::Single, false));
+    let text = task.fmt(config, FormatType::Single, false, false).await?;
+    println!("{}", text);
 
     let options = ["Pick project", "Complete", "Skip", "Delete"]
         .iter()